FEATURE/Add: Add support for heterogeneous model creation
This feature introduces the ability for users to create
heterogeneous models. Users can now specify how individual
components within a system are modeled. For example, in
the case of a cipher with six components, the user can
define that components 1, 2, and 3 will be modeled using the
regular SAT XOR differential model, while components 4, 5,
and 6 will be modeled with the SAT truncated deterministic
model. This provides flexibility for customizing the modeling
approach for different parts of a cipher.