FEATURE/Add: SharedTruncatedDifferencePairedInputDifferentialLinearModel
This model is specifically designed to search for trails satisfying the following condition:

\lambda(f(x) \oplus f(x + a) \oplus f(y) \oplus f(y + a)) = 0

where :math:`\lambda(.)` represents the application of a linear mask and the computation of the parity (i.e., inner product).

The model split the cipher into two parts a high-order truncated differential part and a linear mask.